refactor: change file names and add destructuring
