+++ src/auth/strategy/at.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-jwt';
 
-import cookieExtractor from '../../common/extractors/cookie.extractor';
+import cookieExtractor from '../../common/extractors/cookie';
 
 type JwtPayload = {
   sub: string;
@@ -10,11 +10,11 @@ type JwtPayload = {
 };
 
 @Injectable()
-export class RtStrategy extends PassportStrategy(Strategy, 'jwt-refresh') {
+export class AtStrategy extends PassportStrategy(Strategy, 'jwt') {
   constructor() {
     super({
-      jwtFromRequest: cookieExtractor('refreshToken'),
-      secretOrKey: 'rt-secret',
+      jwtFromRequest: cookieExtractor('accessToken'),
+      secretOrKey: 'at-secret',
     });
   }
