+++ src/booking/module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { BookingService } from './booking.service';
-import { BookingController } from './booking.controller';
+import { BookingService } from './service';
+import { BookingController } from './controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Booking } from './entities/booking.entity';
-import { UserModule } from 'src/user/user.module';
+import { Booking } from './entities/booking';
+import { UserModule } from 'src/user/module';
 
 @Module({
   imports: [UserModule, TypeOrmModule.forFeature([Booking])],
