+++ src/booking/service.ts
@@ -2,10 +2,10 @@ import { Injectable, ForbiddenException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Between, Repository } from 'typeorm';
 
-import { UserService } from 'src/user/user.service';
-import { Room } from 'src/room/entities/room.entity';
+import { UserService } from 'src/user/service';
+import { Room } from 'src/room/entities/room.en';
 
-import { Booking } from './entities/booking.entity';
+import { Booking } from './entities/booking';
 
 @Injectable()
 export class BookingService {
