+++ src/common/guards/admin.ts
@@ -1,5 +1,5 @@
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
-import { UserService } from 'src/user/user.service';
+import { UserService } from 'src/user/service';
 
 @Injectable()
 export class AdminGuard implements CanActivate {
