--- src/hotel/dto/get-hotel.res.dto.ts
@@ -1,6 +1,0 @@
-import { Review } from 'src/review/entities/review.entity';
-import { Hotel } from '../entities/hotel.entity';
-
-export class GetHotelResDto extends Hotel {
-  reviews?: Review[];
-}
