+++ src/hotel/dto/get-hotel.res.ts
@@ -1,0 +1,6 @@
+import { Review } from 'src/review/entities/review';
+import { Hotel } from '../entities/hotel';
+
+export class GetHotelResDto extends Hotel {
+  reviews: Review[];
+}
