+++ src/hotel/dto/get-hotels.query.ts
