+++ src/review/module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { ReviewService } from './review.service';
-import { ReviewController } from './review.controller';
-import { Review } from './entities/review.entity';
-import { UserModule } from 'src/user/user.module';
-import { RoomModule } from 'src/room/room.module';
+import { ReviewService } from './service';
+import { ReviewController } from './controller';
+import { Review } from './entities/review';
+import { UserModule } from 'src/user/module';
+import { RoomModule } from 'src/room/module';
 
 @Module({
   imports: [UserModule, RoomModule, TypeOrmModule.forFeature([Review])],
