+++ src/room/controller.ts
@@ -1,5 +1,5 @@
 import { Controller } from '@nestjs/common';
-import { RoomService } from './room.service';
+import { RoomService } from './service';
 
 @Controller('/rooms')
 export class RoomController {
