+++ src/user/entities/user.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { Booking } from 'src/booking/entities/booking.entity';
-import { Review } from 'src/review/entities/review.entity';
+import { Booking } from 'src/booking/entities/booking';
+import { Review } from 'src/review/entities/review';
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity()
